+++ lib/Vice.Subtitles.ex
@@ -1,6 +1,6 @@
 # File: Vice.Subtitles.ex
 # This file was generated from vice_subtitles.beam
-# Using rebar3_elixir (https://github.com/botsunit/rebar3_elixir)
+# Using rebar3_elixir (https://github.com/G-Corp/rebar3_elixir)
 # MODIFY IT AT YOUR OWN RISK AND ONLY IF YOU KNOW WHAT YOU ARE DOING!
 defmodule Vice.Subtitles do
   def unquote(:"parse")(arg1) do