+++ lib/Vice.Thumbnails.ex
@@ -1,0 +1,12 @@
+# File: Vice.Thumbnails.ex
+# This file was generated from vice_thumbnails.beam
+# Using rebar3_elixir (https://github.com/G-Corp/rebar3_elixir)
+# MODIFY IT AT YOUR OWN RISK AND ONLY IF YOU KNOW WHAT YOU ARE DOING!
+defmodule Vice.Thumbnails do
+  def unquote(:"generate")(arg1, arg2, arg3) do
+    :erlang.apply(:"vice_thumbnails", :"generate", [arg1, arg2, arg3])
+  end
+  def unquote(:"generate")(arg1, arg2) do
+    :erlang.apply(:"vice_thumbnails", :"generate", [arg1, arg2])
+  end
+end