+++ src/vice.erl
@@ -11,6 +11,7 @@
          , infos/1
          , infos/2
          , info/2
+         , info/3
          , status/1
          , stop/1
          , convert/2
@@ -18,16 +19,12 @@
          , convert/4
         ]).
 
--export([thumbnails_finalize/2]).
-
 %% Video API
 -export([
          screenshot/2
-         , thumbnails/2
-         , thumbnails/3
         ]).
 
-%% DEPRECATED
+%% DEPRECATED -- REMOVE  in 1.0.0
 -export([
          webvtt/2
          , webvtt/3
@@ -37,6 +34,8 @@
          , to_html5_mp4/3
          , to_html5_ogg/2
          , to_html5_ogg/3
+         , thumbnails/2
+         , thumbnails/3
         ]).
 
 %% gen_server callbacks
@@ -49,7 +48,13 @@
 
 -define(SERVER, ?MODULE).
 
--type info_options() :: [{labels, atom | binary}].
+-type info_option() :: {labels, atom | binary}
+                       | {type, image | audio | video}
+                       | {allowed_extensions, string()}.
+-type info_options() :: [info_option()]
+                        | #{labels => atom | binary,
+                            type => image | audio | video,
+                            allowed_extensions => string()}.
 
 % @doc
 % Start vice application
@@ -63,15 +68,21 @@ start_link() ->
   gen_server:start_link({local, ?SERVER}, ?MODULE, [], []).
 
 % @doc
-% Return the file type (image or video)
+% Return the file type (image, audio or video)
 % @end
 -spec type(File :: file:filename_all()) -> video | image | unknow.
 type(File) ->
   case bucmime:exploded(File) of
-    {<<"image">>, _} -> image;
-    {<<"video">>, _} -> video;
-    {<<"audio">>, _} -> audio;
+    {Type, _} -> bucs:to_atom(Type);
     _ -> unknow
+  end.
+
+type(File, Options) ->
+  case lists:keyfind(type, 1, Options) of
+    {type, Type} ->
+      Type;
+    false ->
+      type(File)
   end.
 
 % @equiv infos(File, [])
@@ -80,18 +91,25 @@ infos(File) ->
   infos(File, []).
 
 % @doc
-% Return the media informations
+% Return all media informations
 % @end
 -spec infos(File :: file:filename_all(), Options :: info_options()) -> {ok, term()} | {error, term()}.
-infos(File, Options) ->
-  gen_server:call(?SERVER, {infos, File, Options}).
+infos(File, Options) when is_list(Options) ->
+  gen_server:call(?SERVER, {infos, File, Options});
+infos(File, Options) when is_map(Options) ->
+  infos(File, maps:to_list(Options)).
 
-% @doc
-% Return the given media informations
-% @end
+% @equiv info(File, Info, [])
 -spec info(File :: file:filename_all(), Info :: atom()) -> {ok, term()} | {error, term()}.
 info(File, Info) ->
-  gen_server:call(?SERVER, {info, File, Info}).
+  info(File, Info, []).
+
+% @doc
+% Return the given media information
+% @end
+-spec info(File :: file:filename_all(), Info :: atom(), Options :: info_options()) -> {ok, term()} | {error, term()}.
+info(File, Info, Options) ->
+  gen_server:call(?SERVER, {info, File, Info, Options}).
 
 % @doc
 % Return the conversion status
@@ -174,111 +192,22 @@ screenshot(Movie, Out) ->
       {error, invalid_media}
   end.
 
-% @deprecated use thumbnails/2
-webvtt(Movie, OutName) -> thumbnails(Movie, OutName).
-
-% @deprecated use thumbnails/3
-webvtt(Movie, OutName, Options) -> thumbnails(Movie, OutName, Options).
-
-% @equiv thumbnails(Movie, OutName, [{every, 1}, {width, 100}, {out_path, "."}, {sprite, true}, {assets_path, ""}])
-thumbnails(Movie, OutName) ->
-  thumbnails(Movie, OutName, [{every, 1}, {width, 100}, {out_path, "."}, {sprite, true}, {assets_path, ""}]).
-
-% @doc
-% Generate a video thumbnails (.vtt + sprite)
-%
-% Options:
-% <ul>
-% <li><tt>every :: integer()</tt></li>
-% <li><tt>width :: integer()</tt></li>
-% <li><tt>out_path :: string()</tt></li>
-% <li><tt>sprite :: true | false</tt></li>
-% </ul>
-% @end
--spec thumbnails(Movie :: binary() | string(),
-                 OutName :: binary() | string(),
-                 Options :: list()) -> ok | {error, term()}.
-thumbnails(Movie, OutName, Options) ->
-  Every = buclists:keyfind(every, 1, Options, 1),
-  OutPath = buclists:keyfind(out_path, 1, Options, "."),
-  SpritesPath = filename:join(OutPath, OutName),
-  case bucfile:make_dir(SpritesPath) of
-    ok ->
-      SpritesFiles = filename:join(SpritesPath, "%016d.png"),
-      convert(Movie, SpritesFiles, [{output_format, "image2"},
-                                    {video_filtergraph, "fps=1/" ++ bucs:to_string(Every)}],
-              {fun ?MODULE:thumbnails_finalize/2, {Movie, OutName, Options}});
-
-    Error ->
-      Error
-  end.
-
-% @hidden
-thumbnails_finalize({ok, _, _}, {Movie, OutName, Options}) ->
-  Every = buclists:keyfind(every, 1, Options, 1),
-  OutPath = buclists:keyfind(out_path, 1, Options, "."),
-  AssetsPath = buclists:keyfind(assets_path, 1, Options, ""),
-  Width = buclists:keyfind(width, 1, Options, 150),
-  Sprite = buclists:keyfind(width, 1, Options, true),
-  SpritesPath = filename:join(OutPath, OutName),
-  AllSprites = filename:join(SpritesPath, "*.png"),
-  convert(AllSprites, [{geometry, Width}], sync),
-  [F|_] = Sprites = filelib:wildcard(AllSprites),
-  case {vice:infos(F), vice:info(Movie, duration)} of
-    {{ok, #{page_width := Width,
-            page_height := Height,
-            page_x_offset := X,
-            page_y_offset := Y}},
-     {ok, Duration}} ->
-      {Lines, Columns} = vice_utils:tile(length(Sprites)),
-      case generate_vtt(OutName, OutPath, Sprites, AssetsPath, Every, Duration, Lines, Columns, Width, Height, X, Y, Sprite) of
-        ok -> ok;
-        Error ->
-          bucfile:remove_recursive(SpritesPath),
-          Error
-      end;
-    {{ok, _}, Error} ->
-      bucfile:remove_recursive(SpritesPath),
-      Error;
-    {Error, {ok, _}} ->
-      bucfile:remove_recursive(SpritesPath),
-      Error;
-    {Error, _} ->
-      bucfile:remove_recursive(SpritesPath),
-      Error
-  end.
-
-% TODO : remove in 1.0.0
-% @deprecated
-to_html5_webm(Input, Output) ->
-  to_html5_webm(Input, Output, undefined).
-% @deprecated
-to_html5_webm(Input, Output, Fun) ->
-  convert(Input, Output, [{output_format, "webm"}, {vcodec, "libvpx"}, {output_acodec, "libvorbis"}, {output_frame_size, "640x360"}], Fun).
-% @deprecated
-to_html5_mp4(Input, Output) ->
-  to_html5_mp4(Input, Output, undefined).
-% @deprecated
-to_html5_mp4(Input, Output, Fun) ->
-  convert(Input, Output, [{output_format, "mp4"}, {vcodec, "libx264"}, {output_acodec, "libfaac"}, {output_frame_size, "640x360"}], Fun).
-% @deprecated
-to_html5_ogg(Input, Output) ->
-  to_html5_ogg(Input, Output, undefined).
-% @deprecated
-to_html5_ogg(Input, Output, Fun) ->
-  convert(Input, Output, [{vcodec, "libtheora"}, {output_acodec, "libvorbis"}, {output_frame_size, "640x360"}], Fun).
-
 % @hidden
 init(_) ->
-  {ok, #{
-     video => start_encoders(video, ?VIDEO_ENCODERS),
-     image => start_encoders(image, ?PHOTO_ENCODERS),
-     audio => start_encoders(audio, ?AUDIO_ENCODERS)
-    }}.
+  init(
+    buclists:merge_keylists(
+      1,
+      doteki:get_env([vice, encoders], []),
+      maps:to_list(?DEFAULT_ENCODERS)),
+    #{}).
+
+init([], Acc) -> {ok, Acc};
+init([{Type, _}|Rest], Acc) ->
+  init(Rest, maps:put(Type, start_encoders(Type), Acc)).
 
 % @hidden
 handle_call({infos, File, Options}, _From, State) ->
-  case get_encoder(File, State, false) of
+  case get_encoder(File, State, false, Options) of
     {false, {ok, Encoder}} ->
       Reply = gen_server:call(Encoder, {infos, File, Options}),
       release_encoder(Encoder),
@@ -289,10 +218,10 @@ handle_call({infos, File, Options}, _From, State) ->
     Error ->
       {reply, Error, State}
   end;
-handle_call({info, File, Info}, _From, State) ->
-  case get_encoder(File, State, false) of
+handle_call({info, File, Info, Options}, _From, State) ->
+  case get_encoder(File, State, false, Options) of
     {false, {ok, Encoder}} ->
-      Reply = gen_server:call(Encoder, {info, File, Info}),
+      Reply = gen_server:call(Encoder, {info, File, Info, Options}),
       release_encoder(Encoder),
       {reply, Reply, State};
     {true, {ok, Encoder}} ->
@@ -302,7 +231,7 @@ handle_call({info, File, Info}, _From, State) ->
       {reply, Error, State}
   end;
 handle_call({convert, In, Out, Options, Fun}, From, State) ->
-  case get_encoder(In, State, false) of
+  case get_encoder(In, State, false, Options) of
     {Multi, {ok, Encoder}} ->
       gen_server:cast(Encoder, {convert, In, Out, Options, Multi, Fun, From}),
       {noreply, State};
@@ -353,24 +282,31 @@ code_change(_OldVsn, State, _Extra) ->
 
 % Private
 
-start_encoders(Type, Default) ->
-  lager:debug("Start ~p encoders...", [Type]),
-  case poolgirl:add_pool(Type, {vice_prv_encoder,
-                                 start_link,
-                                 [Type, doteki:get_env([vice, encoders, Type], Default)]}) of
-    {ok, N} ->
-      lager:debug("~p ~p encoder started!", [N, Type]),
-      true;
-    {error, Reason} ->
-      lager:error("Faild to start ~p encoder: ~p", [Type, Reason]),
-      false
+start_encoders(Type) ->
+  lager:debug("Start ~p encoder...", [Type]),
+  case doteki:get_env([vice, encoders, Type], maps:get(Type, ?DEFAULT_ENCODERS, undefined)) of
+    undefined ->
+          lager:error("Encoder for ~p undefined.", [Type]),
+          false;
+    Encoders ->
+      case poolgirl:add_pool(Type, {vice_prv_encoder,
+                                    start_link,
+                                    [Type, Encoders]},
+                             #{allow_empty_pool => false}) of
+        {ok, N} ->
+          lager:debug("~p ~p encoder started!", [N, Type]),
+          true;
+        {error, Reason} ->
+          lager:error("Faild to start ~p encoder: ~p", [Type, Reason]),
+          false
+      end
   end.
 
-get_encoder([File|_], State, _) when is_list(File);
-                                     is_binary(File) ->
-  get_encoder(File, State, true);
-get_encoder(File, State, Type) ->
-  case type(File) of
+get_encoder([File|_], State, _, Options) when is_list(File);
+                                              is_binary(File) ->
+  get_encoder(File, State, true, Options);
+get_encoder(File, State, Type, Options) ->
+  case type(File, Options) of
     unknow ->
       {error, invalid_file};
     EncoderType ->
@@ -378,7 +314,7 @@ get_encoder(File, State, Type) ->
         #{EncoderType := true} ->
           {Type, poolgirl:checkout(EncoderType)};
         _ ->
-          {error, invalid_file}
+          {error, {encoder_not_found, EncoderType}}
       end
   end.
 
@@ -386,53 +322,38 @@ release_encoder(?MODULE) -> ok;
 release_encoder(Encoder) ->
   poolgirl:checkin(Encoder).
 
-generate_vtt(OutName, OutPath, AllSprites, AssetsPath, Every, Duration, _Lines, _Columns, Width, Height, _X, _Y, false) ->
-  VttFile = filename:join(OutPath, OutName ++ ".vtt"),
-  case file:open(VttFile, [write]) of
-    {ok, IO} ->
-      io:format(IO, "WEBVTT~n", []),
-      vvtline(AllSprites, 0, Every, Duration, filename:join(AssetsPath, OutName), Width, Height, IO),
-      file:close(IO);
-    Error ->
-      Error
-  end;
+% DEPRECATED FUNTIONS --------------------------------------------------------------------------------------------------------
+% TODO : remove in 1.0.0
 
-generate_vtt(OutName, OutPath, AllSprites, AssetsPath, Every, Duration, Lines, Columns, Width, Height, X, Y, true) ->
-  convert(AllSprites, filename:join(OutPath, OutName ++ ".png"), [{tile, Columns, Lines}, {geometry, Width, Height, X, Y}], sync),
-  VttFile = filename:join(OutPath, OutName ++ ".vtt"),
-  SpritesPath = filename:join(OutPath, OutName),
-  bucfile:remove_recursive(SpritesPath),
-  case file:open(VttFile, [write]) of
-    {ok, IO} ->
-      io:format(IO, "WEBVTT~n", []),
-      vvtline(length(AllSprites), 0, 0, 0, Every, Duration, filename:join(AssetsPath, OutName ++ ".png"), Width, Height, Lines, Columns, IO),
-      file:close(IO);
-    Error ->
-      Error
-  end.
+% @deprecated use vice_thumbnails:generate/2
+webvtt(Movie, OutName) -> thumbnails(Movie, OutName).
 
-vvtline(0, _, _, _, _, _, _, _, _, _, _, _) ->
-  ok;
-vvtline(N, L, C, Start, Every, Duration, SpriteFile, Width, Height, Lines, Columns, IO) ->
-  End = Start + Every,
-  X = C * Width,
-  Y = L * Height,
-  io:format(IO, "~n~s --> ~s~n", [vice_utils:to_full_hms(Start), vice_utils:to_full_hms(End)]),
-  io:format(IO, "~s#xywh=~p,~p,~p,~p~n", [SpriteFile, X, Y, Width, Height]),
-  {NL, NC} = new_position(L, C, Lines, Columns),
-  vvtline(N - 1, NL, NC, End, Every, Duration, SpriteFile, Width, Height, Lines, Columns, IO).
+% @deprecated use vice_thumbnails:generate/3
+webvtt(Movie, OutName, Options) -> thumbnails(Movie, OutName, Options).
 
-new_position(L, C, _Lines, Columns) when C == Columns ->
-  {L + 1, 0};
-new_position(L, C, _Lines, _Columns) ->
-  {L, C + 1}.
+% @deprecated use vice_thumbnails:generate/2
+thumbnails(Movie, OutName) ->
+  vice_thumbnails:generate(Movie, OutName).
 
-vvtline([Sprite], Start, _Every, Duration, AssetsPath, Width, Height, IO) ->
-  io:format(IO, "~n~s --> ~s~n", [vice_utils:to_full_hms(Start), vice_utils:to_full_hms(Duration)]),
-  io:format(IO, "~s/~s#xywh=0,0,~p,~p~n", [AssetsPath, filename:basename(Sprite), Width, Height]);
-vvtline([Sprite|AllSprites], Start, Every, Duration, AssetsPath, Width, Height, IO) ->
-  End = Start + Every,
-  io:format(IO, "~n~s --> ~s~n", [vice_utils:to_full_hms(Start), vice_utils:to_full_hms(End)]),
-  io:format(IO, "~s/~s#xywh=0,0,~p,~p~n", [AssetsPath, filename:basename(Sprite), Width, Height]),
-  vvtline(AllSprites, End, Every, Duration, AssetsPath, Width, Height, IO).
+% @deprecated use vice_thumbnails:thumbnails/3
+thumbnails(Movie, OutName, Options) ->
+  vice_thumbnails:generate(Movie, OutName, Options).
 
+% @deprecated use convert/4 with <tt>html5webm</tt> preset
+to_html5_webm(Input, Output) ->
+  to_html5_webm(Input, Output, undefined).
+% @deprecated use convert/4 with <tt>html5webm</tt> preset
+to_html5_webm(Input, Output, Fun) ->
+  convert(Input, Output, [{output_format, "webm"}, {vcodec, "libvpx"}, {output_acodec, "libvorbis"}, {output_frame_size, "640x360"}], Fun).
+% @deprecated use convert/4 with <tt>html5mp4</tt> preset
+to_html5_mp4(Input, Output) ->
+  to_html5_mp4(Input, Output, undefined).
+% @deprecated use convert/4 with <tt>html5mp4</tt> preset
+to_html5_mp4(Input, Output, Fun) ->
+  convert(Input, Output, [{output_format, "mp4"}, {vcodec, "libx264"}, {output_acodec, "libfaac"}, {output_frame_size, "640x360"}], Fun).
+% @deprecated use convert/4 with <tt>html5ogg</tt> preset
+to_html5_ogg(Input, Output) ->
+  to_html5_ogg(Input, Output, undefined).
+% @deprecated use convert/4 with <tt>html5ogg</tt> preset
+to_html5_ogg(Input, Output, Fun) ->
+  convert(Input, Output, [{vcodec, "libtheora"}, {output_acodec, "libvorbis"}, {output_frame_size, "640x360"}], Fun).