+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [Unreleased](https://github.com/G-Corp/vice/tree/HEAD)
+
+[Full Changelog](https://github.com/G-Corp/vice/compare/0.0.2...HEAD)
+
+**Implemented enhancements:**
+
+- Add HLS options for ffmpeg [\#4](https://github.com/G-Corp/vice/issues/4)
+- Add sox support for audio [\#5](https://github.com/G-Corp/vice/pull/5) ([glejeune](https://github.com/glejeune))
+
+**Closed issues:**
+
+- Convert: passing image content instead of its path [\#7](https://github.com/G-Corp/vice/issues/7)
+
 ## [0.0.2](https://github.com/G-Corp/vice/tree/0.0.2) (2017-08-18)
 [Full Changelog](https://github.com/G-Corp/vice/compare/0.0.1...0.0.2)